+++ tests/_util.py
@@ -3,9 +3,10 @@
 
 __all__ = ('issubtype', 'issequence', 'isiterable', 'remove_all', 'pushd')
 
-from typing import Any, Sequence, Iterable, Tuple, List, Dict
+from typing import Any, Sequence, Iterable, List
 from collections import abc
 import contextlib
+String = str
 
 
 def issubtype(x: Any, t: Any) -> bool:
@@ -13,16 +14,16 @@ def issubtype(x: Any, t: Any) -> bool:
 
 
 def issequence(x: Any) -> bool:
-    return (isinstance(x, (Sequence, abc.Sequence)) and
-            not isinstance(x, (bytes, str)))
+    return (isinstance(x, (Sequence, abc.Sequence))
+            and not isinstance(x, (bytes, str, String)))
 
 
 def isiterable(x: Any) -> bool:
-    return (isinstance(x, (Iterable, abc.Iterable)) and
-            not isinstance(x, (bytes, str, String)))
+    return (isinstance(x, (Iterable, abc.Iterable))
+            and not isinstance(x, (bytes, str, String)))
 
 
-def remove_all(list: List, value: Any) -> None:
+def remove_all(list: List, value: Any) -> None:  # noqa: A002
     list[:] = (item for item in list if item != value)