+++ tests/data/py-chocolatey.test_template/tools/chocolateyInstall.ps1
@@ -1,0 +1,3 @@
+﻿$ErrorActionPreference = 'Stop'; # stop on all errors
+
+Write-Host "I am from [[Country]]"